+++ src/extensions/building/buildingext_hooks.cpp
@@ -37,6 +37,7 @@
 #include "buildingtypeext.h"
 #include "unit.h";
 #include "unitext.h"
+#include "factory.h"
 #include "technotype.h"
 #include "technotypeext.h"
 #include "aircraft.h"
@@ -67,6 +68,7 @@
 
 #include "hooker.h"
 #include "hooker_macros.h"
+#include "spawner.h"
 
 
 /**
@@ -84,6 +86,7 @@ public:
     const InfantryTypeClass* _Crew_Type() const;
     int _How_Many_Survivors() const;
     int _Shape_Number() const;
+    void _Draw_Overlays(Point2D& coord, Rect& rect);
 };
 
 
@@ -270,6 +273,118 @@ int BuildingClassExt::_Shape_Number() const
             shapenum += largest;
         }
     }
+    }
+}
+
+
+/**
+ *  Reimplements the BuildingClass::Draw_Overlays function.
+ *
+ *  @author: ZivDero
+ */
+void BuildingClassExt::_Draw_Overlays(Point2D& coord, Rect& rect)
+{
+    if (BState != BSTATE_CONSTRUCTION)
+    {
+        /**
+         *  Draw the repair animation.
+         */
+        if (IsRepairing)
+        {
+            if (!Map.Is_Shrouded(Center_Coord()) && !(Scen->SpecialFlags.IsFogOfWar || IsFogged) && Visual_Character() != VISUAL_HIDDEN)
+            {
+                Point2D xy = coord;
+                if (!IsPowerOn)
+                    xy -= Point2D(5, 5);
+
+                int delay = Options.Normalize_Delay(14) / 4;
+                if (delay < 2)
+                    delay = 2;
+
+                CC_Draw_Shape(LogicSurface, MouseDrawer, WrenchShape, 6 * (Frame % delay) / (delay - 1), &xy, &rect, SHAPE_ALPHA | SHAPE_WIN_REL | SHAPE_CENTER);
+            }
+        }
+
+        /**
+         *  Draw the power off animation.
+         */
+        if (!IsPowerOn && House->Is_Player_Control())
+        {
+            if (!Map.Is_Shrouded(Center_Coord()) && !(Scen->SpecialFlags.IsFogOfWar || IsFogged))
+            {
+                Point2D xy = coord;
+                if (IsRepairing)
+                    xy += Point2D(10, 10);
+
+                int delay = Options.Normalize_Delay(14) / 4;
+                if (delay < 2)
+                    delay = 2;
+
+                CC_Draw_Shape(LogicSurface, MouseDrawer, PowerOffShape, 6 * (Frame % delay) / (delay - 1), &xy, &rect, SHAPE_ALPHA | SHAPE_WIN_REL | SHAPE_CENTER);
+            }
+        }
+
+        if (IsSelected)
+        {
+            /**
+             *  Draw the primary factory pip.
+             */
+            if (House->Is_Ally(PlayerPtr) || SpiedBy & (1 << (PlayerPtr->Class->House)) || (Spawner::Active && Spawner::Get_Config()->Houses[PlayerPtr->Get_Heap_ID()].IsSpectator))
+            {
+                Point2D xy(coord.X - 10, coord.Y + 10);
+                entry_338(xy, coord, rect);
+            }
+
+            /**
+             *  If this is a factory, and the player has spied its owner, draw the cameo of what it's currently producing.
+             */
+            if (SpiedBy & (1 << (PlayerPtr->Class->House)) || (Spawner::Active && Spawner::Get_Config()->Houses[PlayerPtr->Get_Heap_ID()].IsSpectator))
+            {
+                FactoryClass* factory = House->Is_Human_Control() ? House->Fetch_Factory(Class->ToBuild) : Factory;
+                if (factory != nullptr)
+                {
+                    ObjectClass* obj = factory->Get_Object();
+                    if (obj != nullptr)
+                    {
+                        /**
+                         *  #issue-487
+                         *
+                         *  Adds support for PCX/PNG cameo icons.
+                         *
+                         *  @author: CCHyper
+                         */
+                        const auto technotypeext = Extension::Fetch<TechnoTypeClassExtension>(obj->Techno_Type_Class());
+                        if (technotypeext->CameoImageSurface)
+                        {
+                            /**
+                             *  Draw the cameo pcx image.
+                             */
+                            Rect pcxrect;
+                            pcxrect.X = rect.X + coord.X;
+                            pcxrect.Y = rect.Y + coord.Y;
+                            pcxrect.Width = technotypeext->CameoImageSurface->Get_Width();
+                            pcxrect.Height = technotypeext->CameoImageSurface->Get_Height();
+
+                            SpriteCollection.Draw(pcxrect, *LogicSurface, *technotypeext->CameoImageSurface);
+                        }
+                        else
+                        {
+                            const ShapeFileStruct* shape = obj->Techno_Type_Class()->Get_Cameo_Data();
+
+                            /**
+                             *  Draw the cameo shape.
+                             *
+                             *  Original code used NormalDrawer, which is the old Red Alert shape
+                             *  drawer, so we need to use CameoDrawer here for the correct palette.
+                             */
+                            CC_Draw_Shape(LogicSurface, CameoDrawer, shape, 0, &coord, &rect, SHAPE_ALPHA | SHAPE_WIN_REL | SHAPE_CENTER);
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
 
     return shapenum;
 }
@@ -771,62 +886,6 @@ continue_function:
 }
 
 
-/**
- *  #issue-72
- * 
- *  Fixes the bug where the wrong palette used to draw the cameo of the object
- *  being produced above a enemy spied factory building.
- * 
- *  @author: CCHyper
- */
-DECLARE_PATCH(_BuildingClass_Draw_Spied_Cameo_Palette_Patch)
-{
-    GET_REGISTER_STATIC(TechnoClass *, factory_obj, eax);
-    GET_REGISTER_STATIC(Point2D *, pos_xy, edi);
-    GET_REGISTER_STATIC(Rect *, window_rect, ebp);
-    static TechnoTypeClass *technotype;
-    static TechnoTypeClassExtension *technotypeext;
-    static const ShapeFileStruct *cameo_shape;
-    static Surface *pcx_image;
-    static Rect pcxrect;
-
-    technotype = factory_obj->Techno_Type_Class();
-
-    /**
-     *  #issue-487
-     * 
-     *  Adds support for PCX/PNG cameo icons.
-     * 
-     *  @author: CCHyper
-     */
-    technotypeext = Extension::Fetch<TechnoTypeClassExtension>(technotype);
-    if (technotypeext->CameoImageSurface) {
-
-        /**
-         *  Draw the cameo pcx image.
-         */
-        pcxrect.X = window_rect->X + pos_xy->X;
-        pcxrect.Y = window_rect->Y + pos_xy->Y;
-        pcxrect.Width = technotypeext->CameoImageSurface->Get_Width();
-        pcxrect.Height = technotypeext->CameoImageSurface->Get_Height();
-
-        SpriteCollection.Draw(pcxrect, *LogicSurface, *technotypeext->CameoImageSurface);
-
-    } else {
-
-        cameo_shape = technotype->Get_Cameo_Data();
-
-        /**
-         *  Draw the cameo shape.
-         * 
-         *  Original code used NormalDrawer, which is the old Red Alert shape
-         *  drawer, so we need to use CameoDrawer here for the correct palette.
-         */
-        CC_Draw_Shape(LogicSurface, CameoDrawer, cameo_shape, 0, pos_xy, window_rect, ShapeFlagsType(SHAPE_CENTER|SHAPE_WIN_REL|SHAPE_ALPHA|SHAPE_NORMAL));
-    }
-
-    JMP(0x00428B13);
-}
 
 
 /**
@@ -1060,7 +1119,6 @@ void BuildingClassExtension_Hooks()
      */
     BuildingClassExtension_Init();
 
-    Patch_Jump(0x00428AD3, &_BuildingClass_Draw_Spied_Cameo_Palette_Patch);
     Patch_Jump(0x0042B250, &_BuildingClass_Explode_ShakeScreen_Division_BugFix_Patch);
     Patch_Jump(0x00433BB5, &_BuildingClass_Mission_Open_Gate_Open_Sound_Patch);
     Patch_Jump(0x00433C6F, &_BuildingClass_Mission_Open_Gate_Close_Sound_Patch);
@@ -1084,4 +1142,5 @@ void BuildingClassExtension_Hooks()
     Patch_Jump(0x0042F799, &_BuildingClass_Captured_DontScore_Patch);
     Patch_Jump(0x0042E5F5, &_BuildingClass_Grand_Opening_Assign_FreeUnit_LastDockedBuilding_Patch);
     //Patch_Jump(0x00429220, &BuildingClassExt::_Shape_Number); // It's identical to vanilla, leaving it in in case it's ever needed
+    Patch_Jump(0x00428810, &BuildingClassExt::_Draw_Overlays);
 }