+++ src/new/viniferaevent/viniferaevent_hooks.h
@@ -4,11 +4,11 @@
  *
  *  @project       Vinifera
  *
- *  @file          CNCNET4_GLOBALS.CPP
+ *  @file          VINIFERAEVENT_HOOKS.CPP
  *
- *  @author        CCHyper
+ *  @author        ZivDero
  *
- *  @brief         CnCNet4 global values.
+ *  @brief         Contains the hooks for the Vinifera event class.
  *
  *  @license       Vinifera is free software: you can redistribute it and/or
  *                 modify it under the terms of the GNU General Public License
@@ -25,31 +25,7 @@
  *                 If not, see <http://www.gnu.org/licenses/>.
  *
  ******************************************************************************/
-#include "cncnet4_globals.h"
-#include "cncnet4.h"
+#pragma once
 
 
-/**
- *  Is the CnCNet4 interface active?
- */
-bool CnCNet4::IsEnabled = false;
-
-/**
- *  The host name (Must be running a instance of the dedicated server).
- */
-char CnCNet4::Host[256] = { "server.cncnet.org" };
-unsigned CnCNet4::Port = 9001;
-
-/**
- *  Clients connect to each other rather than the server?
- */
-bool CnCNet4::Peer2Peer = false;
-
-bool CnCNet4::IsDedicated = false;
-
-/**
- *  Use the UDP interface instead of IPX?
- */
-bool CnCNet4::UseUDP = true;
-
-struct sockaddr_in CnCNet4::Server;
+void ViniferaEvent_Hooks();